LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/onpaint-changedElements-expected.txt


FAIL onpaint event should have a list of changed elements assert_equals: onpaint should fire after #childa changes. expected 1 but got 0
